Moves some tests out of plasticc test module and into test modules for new classes
